Start simplifying a bit to avoid using semantic predicates and actions
It's better to simplify the main parser+lexer and instead rely on
simple miniparsers to help with some complicated parts of the spec,
such as attribute and element constructors.

I have an idea of how to do attributes, but elements may require
specifying a *lot* of tokens where the old ElementContentChar was.